feat(plasma-theme-builder): Rename groups of tokens to actual name
