+++ website/plasma-theme-builder/src/builder/createTheme.ts
@@ -7,8 +7,8 @@ import type {
 } from '@salutejs/plasma-tokens-utils';
 
 import {
-    textIconsTokenGetters,
-    controlsSurfacesTokenGetters,
+    textTokenGetters,
+    surfaceTokenGetters,
     backgroundTokenGetters,
     overlayTokenGetters,
 } from './themeTokenGetters';
@@ -92,9 +92,9 @@ const getTokensBackgroundByGroups = <T extends string>(
 
 const getThemeModeTokens = <T extends ThemeMode>(config: ThemeConfig, mode: T): Theme[T] => {
     return {
-        textIcons: getTokensByGroups<TextIconsTokenName>(textIconsTokenGetters, config, mode),
-        controlsSurfaces: getTokensByGroups<ControlsSurfacesName>(controlsSurfacesTokenGetters, config, mode),
-        backgrounds: getTokensBackgroundByGroups<BackgroundName>(backgroundTokenGetters, config, mode),
+        text: getTokensByGroups<TextIconsTokenName>(textTokenGetters, config, mode),
+        surface: getTokensByGroups<ControlsSurfacesName>(surfaceTokenGetters, config, mode),
+        background: getTokensBackgroundByGroups<BackgroundName>(backgroundTokenGetters, config, mode),
         overlay: getTokensByGroups<OverlayName>(overlayTokenGetters, config, mode),
     };
 };
